app/requirements.txt
rumps>=0.4.0
Pillow>=10.0.0
keyring>=24.0.0
requests>=2.28.0
